+++ changelog.md
@@ -1,11 +1,12 @@
 # Retribution v1.5.0
 
 ## Features/Improvements
+* **[Campaigns]** Ability to define invisible FOBs
 * **[Plugins]** Improvements to AI support for EW Script 2.0
 * **[Config]** New preference setting to trigger the first-start window on every start (could help in scenarios multiple Retribution instances need to run concurrently)
 
 ## Fixes
-
+* **[Flight Plans]** Fixed a bug when a package was created with only escort flights
 
 # Retribution v1.4.0